requirements.txt
django
pillow
django-crispy-forms
django-location-field
psycopg2